LPS-168476 Extract PreviewButton
